Unnecessary css class was causing a test failure. Removed class.
The voting round navigation was put in without running the tests. I
searched through the assets, and I see neither css nor js that is
selecting on voting-round, so it feels more semantic. Also looking at
the live site and removing the class from the navigation links through
inspector didn't appear to change anything.

Also, semantically, having this class on both the high-level div and on
the navigation links feels weird.